train planner restructuring
