+++ pyproject.toml
@@ -43,7 +43,7 @@ dependencies = [
     "plotly",
     "typing_extensions",
     "pyyaml",
-    "polars >= 1.19.0",
+    "polars >= 1.20.0",
     "pyarrow",
     "requests",
     "PyYAML==6.0.2",
